+++ frontend/src/app/bookbook/user/lendlist/page.tsx
@@ -2,10 +2,9 @@
 
 import { useState, useEffect } from 'react';
 import { Search, Book } from 'lucide-react';
-import Pagination from '../../components/Pagination';
-import UserSidebar from '../../components/UserSidebar';
+import Pagination from '../../../components/Pagination';
 import LendListCard from './LendListCard';
-import ReviewModal from '../../components/ReviewModal';
+import ReviewModal from '../../../components/ReviewModal';
 import { MyBook, PaginationInfo } from './types';
 import { dummyLendListBooks } from './dummyData';
 
@@ -260,19 +259,14 @@ export default function LendListPage() {
 
   if (loading) {
     return (
-      <div className="flex">
-        <UserSidebar />
-        <div className="flex-1 flex justify-center items-center min-h-screen">
-          <div className="animate-spin rounded-full h-12 w-12 border-b-2 border-gray-900"></div>
-        </div>
+      <div className="flex justify-center items-center min-h-screen">
+        <div className="animate-spin rounded-full h-12 w-12 border-b-2 border-gray-900"></div>
       </div>
     );
   }
 
   return (
-    <div className="flex">
-      <UserSidebar />
-      <div className="flex-1 p-6">
+    <div className="w-full">
         <div className="mb-8">
           <h1 className="text-3xl font-bold mb-2">내가 등록한 도서</h1>
           <p className="text-gray-600">
@@ -361,7 +355,6 @@ export default function LendListPage() {
             onSubmit={submitReview}
           />
         )}
-      </div>
     </div>
   );
 }
