+++ frontend/src/app/bookbook/user/rentlist/page.tsx
@@ -2,10 +2,9 @@
 
 import { useState, useEffect } from 'react';
 import { Search, Book } from 'lucide-react';
-import Pagination from '../../components/Pagination';
-import UserSidebar from '../../components/UserSidebar';
+import Pagination from '../../../components/Pagination';
 import RentListCard from './RentListCard';
-import ReviewModal from '../../components/ReviewModal';
+import ReviewModal from '../../../components/ReviewModal';
 import { RentedBook, PaginationInfo } from './types';
 import { dummyRentedBooks } from './dummyData';
 
@@ -216,19 +215,14 @@ export default function RentListPage() {
 
   if (loading) {
     return (
-      <div className="flex">
-        <UserSidebar />
-        <div className="flex-1 flex justify-center items-center min-h-screen">
-          <div className="animate-spin rounded-full h-12 w-12 border-b-2 border-gray-900"></div>
-        </div>
+      <div className="flex justify-center items-center min-h-screen">
+        <div className="animate-spin rounded-full h-12 w-12 border-b-2 border-gray-900"></div>
       </div>
     );
   }
 
   return (
-    <div className="flex">
-      <UserSidebar />
-      <div className="flex-1 p-6">
+    <div className="w-full">
         <div className="mb-8">
           <h1 className="text-3xl font-bold mb-2">빌린 도서 목록</h1>
           <p className="text-gray-600">
@@ -316,7 +310,6 @@ export default function RentListPage() {
             onSubmit={submitReview}
           />
         )}
-      </div>
     </div>
   );
 }
