+++ frontend/src/app/components/UserSidebar.tsx
@@ -38,9 +38,9 @@ export default function UserSidebar() {
           <div>
             <h3 className="text-lg font-bold text-gray-600 mb-3">내 도서</h3>
             <div className="space-y-1">
-              <SidebarItem href="/bookbook/lendlist" label="도서 등록 내역" />
-              <SidebarItem href="/bookbook/rentlist" label="도서 대여 내역" />
-              <SidebarItem href="/bookbook/wishlist" label="찜한 도서" />
+              <SidebarItem href="/bookbook/user/lendlist" label="도서 등록 내역" />
+              <SidebarItem href="/bookbook/user/rentlist" label="도서 대여 내역" />
+              <SidebarItem href="/bookbook/user/wishlist" label="찜한 도서" />
             </div>
           </div>
